+++ src/App.js
@@ -13,6 +13,7 @@ import Users from './pages/Users';
 import Header from './components/Header';
 import Dashboard from './pages/Dashboard';
 import ProjectEdit from './pages/ProjectEdit';
+import ClientDetail from './pages/ClientDetail';
 
 function App() {
   const LoginContainer = () => (
@@ -34,6 +35,7 @@ function App() {
         <Route exact path='/projects/edit/:id' component={ProjectEdit} />
         <Route exact path='/clients' component={Clients} />
         <Route path='/clients/new' component={NewClient} />
+        <Route exact path='/clients/:id' component={ClientDetail} />
         <Route path='/clients/edit' component={NewClient} />
         <Route path='/profile' component={ProfilePage} />
         <Route exact path='/tasks' component={Tasks} />
