+++ src/components/client/AddClientButton.js
@@ -1,7 +1,14 @@
 import React from 'react';
 import AddButton from '../global/AddButton';
+import {Link} from 'react-router-dom';
 
 export default function AddClientButton() {
   const buttonName = 'New Client';
-  return <AddButton buttonName={buttonName} />;
+  return (
+    <React.Fragment>
+      <Link className='text-light' to='/clients/new'>
+        <AddButton buttonName={buttonName} />
+      </Link>
+    </React.Fragment>
+  );
 }
